requirements.txt
# Core Dependencies
python>=3.9

# LangChain and related
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-deepseek>=0.1.0
langgraph>=0.0.40
langchain-neo4j>=0.1.0

# OpenAI
openai>=1.0.0

# Vector Stores and Embeddings
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# Neo4j
neo4j>=5.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment Management
python-dotenv>=1.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Optional: For development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
